feat: 调整歌词详情页布局
