+++ src/pages/musicDetail/components/content/albumCover/operations.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import {Image, Pressable, StyleSheet, View} from 'react-native';
 import rpx from '@/utils/rpx';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import MusicSheet from '@/core/musicSheet';
 
 import Download from '@/core/download';
 import LocalMusicSheet from '@/core/localMusicSheet';
@@ -14,6 +13,7 @@ import {showPanel} from '@/components/panels/usePanel';
 import TrackPlayer from '@/core/trackPlayer';
 import {iconSizeConst} from '@/constants/uiConst';
 import PersistStatus from '@/core/persistStatus';
+import HeartIcon from '../heartIcon';
 
 export default function Operations() {
     //briefcase-download-outline  briefcase-check-outline checkbox-marked-circle-outline
@@ -23,10 +23,6 @@ export default function Operations() {
 
     const rate = PersistStatus.useValue('music.rate', 100);
     const orientation = useOrientation();
-
-    const favIndex = MusicSheet.useMusicFavIndex(musicItem);
-
-    console.log(favIndex);
 
     return (
         <View
@@ -38,27 +34,7 @@ export default function Operations() {
                       }
                     : null,
             ]}>
-            {favIndex !== -1 ? (
-                <Icon
-                    name="heart"
-                    size={iconSizeConst.normal}
-                    color="red"
-                    onPress={() => {
-                        MusicSheet.removeMusicByIndex('favorite', favIndex);
-                    }}
-                />
-            ) : (
-                <Icon
-                    name="heart-outline"
-                    size={iconSizeConst.normal}
-                    color="white"
-                    onPress={() => {
-                        if (musicItem) {
-                            MusicSheet.addMusic('favorite', musicItem);
-                        }
-                    }}
-                />
-            )}
+            <HeartIcon />
             <Pressable
                 onPress={() => {
                     if (!musicItem) {
